refactor(Contact): move contact form logic to a separate hook and clean up imports
